[DO NOT MERGE] Hacks to bypass constraints from objectives and BG source
These are ONLY for development!! They are needed to be able to use a
Nightscout instance as BG source to test-run a closed-loop operation
with combov2 as the pump driver. These tests are run on special test
pumps that are NOT being used for actual insulin delivery.

DO NOT USE IN PRODUCTION!!!!

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>