combov2+comboctl: Use sequenced dispatcher to avoid race conditions
It is difficult to avoid ComboCtl API calls be made by multiple threads.
Most of the time, these calls still happen sequentially, not concurrently,
but the fact that they _may_ happen concurrently is still a problem.
Enforce sequential execution by using a "sequenced dispatcher", which is
a Kotlin CoroutineDispatcher that is guaranteed to never execute more than
one task at the same time.